BUG: Series[sparse].where losing fill_value

+++ pandas/core/arrays/sparse/array.py
@@ -1355,7 +1355,7 @@ class SparseArray(OpsMixin, PandasObject, ExtensionArray):
         # NB: may not preserve dtype, e.g. result may be Sparse[float64]
         #  while self is Sparse[int64]
         naive_implementation = np.where(mask, self, value)
-        result = type(self)._from_sequence(naive_implementation)
+        result = type(self)._from_sequence(naive_implementation, dtype=self.dtype)
         return result
 
     # ------------------------------------------------------------------------
